SIMPLEX-127 Use refset types on MAIN if none found
Snowstorm allows creating a special empty version of the main codesystem using the 20000101 effectiveTime. This allows creating child codesystems that are completely empty, useful for importing a subontology or an Edition without inheriting any content from MAIN.

The issue is that when exporting RF2 from a child or grandchild codesystem when the 2000 base is used there are no refset types configured. This is because the refset types are saved on the MAIN branch after the empty 2000 base timepoint.

Solution: Update the export function to detect when no refset types are configured and take them from the latest version of MAIN instead.